// src/app/models/Usuario.ts
export class Usuario {
  idUsuario: number = 0;
  username: string = '';
  password: string = '';
  enabled: boolean = false;
  nombres: string = '';
  apellidos: string = '';
  correo: string = '';
  sitioWeb: string = '';
  telefono: number = 0;
  nombreEmpresa: string = '';
  esPremium: boolean = false;
}